Move "CapturedMaterial" component to separate directory
