+++ frontend/src/features/gameplay/CapturedMaterial.tsx
